make dt1 be calculated in single-precision always since its exact value is not that important: if you are truly worried about shift in the 8th digit then you should be more worried about your safety factors

+++ src/astaroth/DSL/handwritten_end.h
@@ -32,7 +32,7 @@
 
 	if (AC_lfirst__mod__cdata)
 	{
-		reduce_max(dt1_max__mod__cdata,AC_dt1_max)
+		reduce_max((float)dt1_max__mod__cdata,AC_dt1_max)
 	}
 #if LCHIRAL
   	if (AC_ixx_chiral__mod__chiral != 0) write(F_XX_CHIRAL,rk_intermediate(F_XX_CHIRAL, DF_XX_CHIRAL,  step_num, AC_dt__mod__cdata) )
